centralized type interface for Hompage
